xds: fix wrong number type for generated raw config (1.31.x backport) (#7884)
JsonUtil is expecting numbers in Map<String, ?> to be Double. The generated raw config (Ma<String, ?>) should use Double for numerical values put into it.

+++ xds/src/main/java/io/grpc/xds/XdsNameResolver.java
@@ -310,10 +310,10 @@ final class XdsNameResolver extends NameResolver {
         rawHeaderMatcherBuilder.put("regexMatch", regexMatch.pattern());
       }
       if (rangeMatch != null) {
-        rawHeaderMatcherBuilder
-            .put(
-                "rangeMatch",
-                ImmutableMap.of("start", rangeMatch.getStart(), "end", rangeMatch.getEnd()));
+        rawHeaderMatcherBuilder.put(
+            "rangeMatch",
+            ImmutableMap.of("start", Long.valueOf(rangeMatch.getStart()).doubleValue(),
+                "end", Long.valueOf(rangeMatch.getEnd()).doubleValue()));
       }
       if (presentMatch != null) {
         rawHeaderMatcherBuilder.put("presentMatch", presentMatch);
